requirements.txt
streamlit==1.39.0
torch==2.5.0
torchvision==0.20.0
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
numpy==2.1.2
Pillow==11.0.0
python-dotenv==1.0.1
google-generativeai==0.8.3
pytesseract==0.3.13
PIL==1.1.6
Image==1.5.33

